+++ src/components/Checkout/CheckoutReceipt.vue
@@ -4,7 +4,7 @@
 			Order Confirmation
 		</h2>
 		<div class="checkout-receipt__wrapper tw-bg-primary tw-p-4 tw-rounded-sm tw-border tw-border-tertiary">
-			<section data-test="lender-info" class="section section--lender-info">
+			<section data-testid="lender-info" class="section section--lender-info">
 				<div>{{ formattedTransactionTime }}</div>
 				<div class="fs-exclude">
 					{{ lender.firstName }} {{ lender.lastName }}
@@ -15,7 +15,7 @@
 			</section>
 			<section>
 				<div
-					data-test="print-kcard-msg"
+					data-testid="print-kcard-msg"
 					class="section tw-text-center hide-for-print"
 					v-if="printableKivaCards.length > 0"
 				>
@@ -29,16 +29,16 @@
 				</div>
 				<ul class="checkout-receipt__item-list">
 					<!-- Loans -->
-					<li
+					<li data-testid="receipt-item-loan"
 						class="section"
 						v-for="loan in loans"
 						:key="loan.id"
 					>
 						<div
-							data-test="loan"
+							data-testid="loan"
 							class="loan"
 						>
-							<h3 class="loan__name fs-exclude">
+							<h3 class="loan__name fs-exclude" data-testid="loan-name">
 								<template v-if="disableRedirects">
 									{{ loan.loan.name }}
 								</template>
@@ -50,7 +50,7 @@
 								</router-link>
 							</h3>
 							<div class="tw-text-small">
-								<p class="tw-text-secondary">
+								<p class="tw-text-secondary" data-testid="loan-country">
 									<template v-if="loan.loan.geocode && loan.loan.geocode.city">
 										{{ loan.loan.geocode.city }},
 									</template>
@@ -58,42 +58,48 @@
 										{{ loan.loan.geocode.country.name }}
 									</template>
 								</p>
-								<p v-if="loan.loan.use">
+								<p v-if="loan.loan.use" data-testid="loan-use">
 									A loan helps {{ loan.loan.use }}
 								</p>
 							</div>
-							<div class="loan__amount tw-text-h3">
+							<div class="loan__amount tw-text-h3" data-testid="loan-amount">
 								${{ loan.price }}
 							</div>
 						</div>
 					</li>
 					<!-- Kiva Cards -->
-					<li
+					<li data-testid="receipt-item-kcard"
 						class="section"
 						v-for="card in kivaCards"
 						:key="card.id"
 					>
 						<div
-							data-test="kcard"
+							data-testid="kcard"
 							class="loan kcard"
 						>
 							<template v-if="card.kivaCardObject.deliveryType === 'print'">
 								<div>
-									<h3 class="loan__name loan__name--inline">
+									<h3 class="loan__name loan__name--inline"
+										data-testid="kcard-type-print"
+									>
 										Print-it-yourself Kiva Card
 									</h3>
 									<kv-icon
+										data-testid="kcard-tool-tip-control"
 										name="question"
 										:id="`print-card-${card.id}`"
 										class="loan__question-icon tw-fill-current tw-text-tertiary"
 									/>
-									<kv-tooltip :controller="`print-card-${card.id}`" theme="mint">
+									<kv-tooltip data-testid="kcard-tool-tip-text"
+										:controller="`print-card-${card.id}`" theme="mint"
+									>
 										You can print this card now. We'll also send it to
 										you in an email so you can print it later.
 									</kv-tooltip>
 								</div>
 
 								<kv-button
+									data-testid="kcard-print-button"
 									class="tw-mb-2"
 									target="_blank"
 									:href="`/gifts/kiva-cards/print?giftCode=${card.kivaCardObject.redemptionCode}`"
@@ -102,10 +108,10 @@
 								</kv-button>
 							</template>
 							<template v-else-if="card.kivaCardObject.deliveryType === 'postal'">
-								<h3 class="loan__name">
+								<h3 class="loan__name" data-testid="kcard-type-postal">
 									Postal delivery Kiva Card
 								</h3>
-								<div class="loan__details fs-exclude">
+								<div class="loan__details fs-exclude" data-testid="kcard-details">
 									For: {{ card.kivaCardObject.mailingInfo.firstName }}
 									{{ card.kivaCardObject.mailingInfo.lastName }}<br>
 									{{ card.kivaCardObject.mailingInfo.address }}<br>
@@ -118,10 +124,11 @@
 								</div>
 							</template>
 							<template v-else-if="card.kivaCardObject.deliveryType === 'lender'">
-								<h3 class="loan__name">
+								<h3 class="loan__name" data-testid="kcard-type-lender">
 									Kiva Card
 								</h3>
 								<div
+									data-testid="kcard-details"
 									class="loan__details fs-exclude"
 									v-if="card.kivaCardObject.recipient.name"
 								>
@@ -129,10 +136,10 @@
 								</div>
 							</template>
 							<template v-else-if="card.kivaCardObject.deliveryType === 'email'">
-								<h3 class="loan__name">
+								<h3 class="loan__name" data-testid="kcard-type-email">
 									Email delivery Kiva Card
 								</h3>
-								<div class="loan__details fs-exclude">
+								<div class="loan__details fs-exclude" data-testid="kcard-details">
 									For:
 									<template v-if="card.kivaCardObject.recipient.name">
 										{{ card.kivaCardObject.recipient.name }} &ndash;
@@ -140,13 +147,13 @@
 									{{ card.kivaCardObject.recipient.email }}
 								</div>
 							</template>
-							<div class="loan__amount tw-text-h3">
+							<div class="loan__amount tw-text-h3" data-testid="kcard-amount">
 								${{ card.price }}
 							</div>
 						</div>
 					</li>
 					<li
-						data-test="kcard-portfolio"
+						data-testid="kcard-portfolio"
 						class="section tw-text-center"
 						v-if="kivaCards.length > 0"
 					>
@@ -164,31 +171,34 @@
 					<li class="section">
 						<div
 							class="loan"
-							data-test="donation"
+							data-testid="donation"
 						>
 							<h3 class="loan__name">
 								Donation to Kiva
 							</h3>
 							<router-link
 								v-if="receipt.totals.donationTotal > 0"
+								data-testid="print-donation-information"
 								class="smallest"
 								to="/portfolio/donations"
 							>
 								Print Donation Information
 							</router-link>
-							<div class="loan__amount tw-text-h3">
+							<div class="loan__amount tw-text-h3" data-testid="donation-amount">
 								${{ receipt.totals.donationTotal }}
 							</div>
 						</div>
 					</li>
 					<li
-						data-test="receipt-total"
+						data-testid="receipt-total"
 						class="section total"
 					>
 						<h3 class="total__header tw-text-h3">
 							Total:
 						</h3>
-						<span class="total__amount tw-text-h3">${{ receipt.totals.itemTotal }}</span>
+						<span class="total__amount tw-text-h3"
+							data-testid="total-amount"
+						>${{ receipt.totals.itemTotal }}</span>
 					</li>
 				</ul>
 			</section>
@@ -198,65 +208,70 @@
 				</h3>
 				<ul class="payments__list">
 					<li
-						data-test="kcard-payment"
+						data-testid="payment-type-kcard"
 						class="total"
 						v-if="receipt.totals.redemptionCodeAppliedTotal > 0"
 					>
 						<span class="total__header tw-text-h3">Kiva Card:</span>
-						<span class="total__amount tw-text-h3">
+						<span class="total__amount tw-text-h3" data-testid="kcard-payment-amount">
 							${{ receipt.totals.redemptionCodeAppliedTotal }}
 						</span>
 					</li>
 					<li
-						data-test="free-trial"
+						data-testid="payment-type-free-trial"
 						class="total"
 						v-if="receipt.totals.freeTrialAppliedTotal > 0"
 					>
 						<span class="total__header tw-text-h3">Free Trial:</span>
-						<span class="total__amount tw-text-h3">Free!</span>
+						<span class="total__amount tw-text-h3" data-testid="free-trial-payment-amount">Free!</span>
 					</li>
 					<li
-						data-test="free-credit"
+						data-testid="payment-type-free-credit"
 						class="total"
 						v-if="receipt.totals.bonusAppliedTotal > 0"
 					>
 						<span class="total__header tw-text-h3">Free credit:</span>
-						<span class="total__amount tw-text-h3">{{ receipt.totals.bonusAppliedTotal }}</span>
+						<span class="total__amount tw-text-h3"
+							data-testid="free-credit-payment-amount"
+						>{{ receipt.totals.bonusAppliedTotal }}</span>
 					</li>
 					<li
-						data-test="kiva-credit-added"
+						data-testid="payment-type-kiva-credit-added"
 						class="total"
 						v-if="receipt.totals.depositTotals.kivaCreditAdded > 0"
 					>
 						<span class="total__header tw-text-h3">Kiva credit added:</span>
-						<span class="total__amount tw-text-h3">
+						<span class="total__amount tw-text-h3" data-testid="kiva-credit-added-payment-amount">
 							${{ receipt.totals.depositTotals.kivaCreditAdded }}
 						</span>
 					</li>
 					<li
-						data-test="kiva-credit-used"
+						data-testid="payment-kiva-credit-used"
 						class="total"
 						v-if="receipt.totals.depositTotals.kivaCreditUsed > 0"
 					>
 						<span class="total__header tw-text-h3">Kiva credit:</span>
-						<span class="total__amount tw-text-h3">
+						<span class="total__amount tw-text-h3" data-testid="kiva-credit-use-payment-amount">
 							${{ receipt.totals.depositTotals.kivaCreditUsed }}
 						</span>
 					</li>
 					<li
-						data-test="amount-charged"
+						data-testid="payment-type-amount-charged"
 						class="total"
 						v-if="parseFloat(receipt.totals.depositTotals.depositTotal) > 0"
 					>
 						<span class="total__header tw-text-h3">Amount charged:</span>
-						<span class="total__amount tw-text-h3">
+						<span class="total__amount tw-text-h3" data-testid="amount-charged-payment-amount">
 							${{ receipt.totals.depositTotals.depositTotal }}
 						</span>
 					</li>
 				</ul>
 			</section>
 			<section class="section section--print hide-for-print">
-				<button class="print tw-text-link tw-flex tw-items-center tw-gap-1" @click="printReceipt">
+				<button data-testid="print-receipt"
+					class="print tw-text-link tw-flex tw-items-center tw-gap-1"
+					@click="printReceipt"
+				>
 					<kv-icon name="print" class="print__icon tw-fill-current" />
 					<span>Print this receipt</span>
 				</button>