+++ src/components/Checkout/OrderTotals.vue
@@ -1,20 +1,23 @@
 <template>
 	<div class="order-totals" data-testid="basket-order-totals-section">
-		<div v-if="showPromoCreditTotal" class="order-total" data-test="order-total" data-testid="basket-order-total">
+		<div v-if="showPromoCreditTotal" class="order-total" data-testid="basket-order-total">
 			<strong>Order Total: <span class="total-value">{{ itemTotal }}</span></strong>
 		</div>
 
 		<div
 			v-if="showKivaCredit"
 			class="kiva-credit tw-font-medium tw-mb-2"
-			data-test="kiva-credit"
 			data-testid="basket-kiva-credit"
 		>
 			<span v-if="showRemoveKivaCredit">
-				Kiva credit: <span class="total-value">({{ kivaCredit }})</span>
+				Kiva credit: <span data-testid="applied-kiva-credit-amount"
+					class="total-value"
+				>({{ kivaCredit }})</span>
 			</span>
 			<span v-if="showApplyKivaCredit">
-				<del>Kiva credit:</del> <span class="total-value"><del>({{ kivaCredit }})</del></span>
+				<del>Kiva credit:</del> <span data-testid="removed-kiva-credit-amount"
+					class="total-value"
+				><del>({{ kivaCredit }})</del></span>
 			</span>
 			<button
 				v-if="showRemoveKivaCredit"
@@ -40,7 +43,7 @@
 		</div>
 
 		<div v-if="showPromoCreditTotal">
-			<div class="order-total" data-test="promo-total" data-testid="basket-promo-total">
+			<div class="order-total" data-testid="basket-promo-total">
 				<template v-if="availablePromoTotal">
 					{{ availablePromoTotal }}
 				</template>
@@ -51,7 +54,9 @@
 					v-if="promoFundDisplayName"
 				>
 					{{ promoFundDisplayName }}
-				</kv-button> promotion: <span class="total-value">({{ appliedPromoTotal }})</span>
+				</kv-button> promotion: <span data-testid="promo-amount"
+					class="total-value"
+				>({{ appliedPromoTotal }})</span>
 				<button
 					v-if="showRemoveActivePromoCredit"
 					class="remove-credit"
@@ -82,6 +87,7 @@
 			</div>
 			<kv-tooltip
 				class="tooltip"
+				data-testid="promo-tool-tip"
 				controller="promo_name"
 				theme="mint"
 				v-if="promoFundDisplayDescription"
@@ -90,7 +96,7 @@
 			</kv-tooltip>
 		</div>
 
-		<div class="order-total" data-test="total-due">
+		<div class="order-total" data-testid="total-due">
 			<strong>
 				<template v-if="!showPromoCreditTotal">Total: </template>
 				<template v-else>Total Due: </template>