+++ src/components/WwwFrame/SearchBar.vue
@@ -11,6 +11,7 @@
 			id="top-nav-search"
 			ref="input"
 			name="queryString"
+			data-testid="header-search-text-input"
 			:icon="mdiMagnify"
 			:model-value="displayTerm"
 			:can-clear="true"
@@ -47,7 +48,9 @@
 				md:tw-rounded-b
 			"
 		>
-			<li v-for="section in sections" :key="section.name" class="section">
+			<li v-for="section in sections" :key="section.name" class="section"
+				:data-testid="`header-search-results-${section.name}`"
+			>
 				<h2 class="tw-text-base tw-py-0.5">
 					{{ section.name }}
 				</h2>
@@ -56,7 +59,7 @@
 						:key="suggestion.label"
 						@mousedown.prevent
 						@click="runSearch(suggestion)"
-						data-testid="search-bar-result"
+						data-testid="header-search-result-item"
 						class="
 							tw-pl-1.5 tw-py-0.5 tw-rounded-sm
 							tw-font-medium tw-cursor-pointer