+++ CHANGELOG.md
@@ -1,3 +1,117 @@
+## [2.105.0](https://github.com/kiva/ui/compare/v2.104.2...v2.105.0) (2022-04-13)
+
+
+### Features
+
+* kvcontenfulimg component was replace for our kv-components one ([1438e7c](https://github.com/kiva/ui/commit/1438e7cfbccd5fda9554cfd6639ba81ccec34a9f))
+* mg page was updated to have a new style in the signup form ([27d2fc4](https://github.com/kiva/ui/commit/27d2fc47945fe4190b6d4b949b3909d56f6f139e))
+* unused css classes were removed ([1f3cd14](https://github.com/kiva/ui/commit/1f3cd14fc6c5d564d0c33574980d4fc3e5453abf))
+
+### [2.104.2](https://github.com/kiva/ui/compare/v2.104.1...v2.104.2) (2022-04-13)
+
+
+### Bug Fixes
+
+* thanks page layout and upsell for new subscription service users ([98eb2a8](https://github.com/kiva/ui/commit/98eb2a803f86ef9b543fab2bcbc06720e17b7463))
+
+### [2.104.1](https://github.com/kiva/ui/compare/v2.104.0...v2.104.1) (2022-04-12)
+
+
+### Bug Fixes
+
+* loan bundles initialization was updated ([bf33fd6](https://github.com/kiva/ui/commit/bf33fd69cb689768b3a9491feacf4ecb8eaedbdb))
+
+## [2.104.0](https://github.com/kiva/ui/compare/v2.103.1...v2.104.0) (2022-04-11)
+
+
+### Features
+
+* add and update meta info ([0000d82](https://github.com/kiva/ui/commit/0000d821ea4507d1d9d7edf2d29d693fe6587095))
+* add meta information ([432ed48](https://github.com/kiva/ui/commit/432ed48fcb7393c0a08fb532808c47f240b3af26))
+* adding tags to parser filter string logic ([c234d52](https://github.com/kiva/ui/commit/c234d5264c3c8b3742f95676a1df94add077c42d))
+* allowing dots in router and filter parser ([4548376](https://github.com/kiva/ui/commit/454837633f2be8eadfa6ff6d145b8614f9248f7e))
+* avoid unnecessary dot use in regex ([1fd6912](https://github.com/kiva/ui/commit/1fd691235d1f658f067462ab09e6abe39928d864))
+* change tests to get expected result ([883456e](https://github.com/kiva/ui/commit/883456eaa532d7c003e9971a05a5e0c89ab6f366))
+
+
+### Bug Fixes
+
+* adding vmid to override meta description tag ([2678936](https://github.com/kiva/ui/commit/2678936e07e43cd0de4724fb981498ee14cbefa2))
+
+### [2.103.1](https://github.com/kiva/ui/compare/v2.103.0...v2.103.1) (2022-04-11)
+
+
+### Bug Fixes
+
+* added needed testid for header and banner ([242f669](https://github.com/kiva/ui/commit/242f6691cbe228ea2b696666a12bf4bc8b818980))
+* added some top level data-testids for elements with same data-testid ([4bfd733](https://github.com/kiva/ui/commit/4bfd733760f780fc2990b00aa307a399dcf53f1d))
+* fix SearchBar.spec.js test ([20ea3d5](https://github.com/kiva/ui/commit/20ea3d5f3668038b4e6cfab55431fae335f4a0c5))
+
+## [2.103.0](https://github.com/kiva/ui/compare/v2.102.0...v2.103.0) (2022-04-11)
+
+
+### Features
+
+* all the setting for mg_amount_selector experiment were removed ([17d2b5f](https://github.com/kiva/ui/commit/17d2b5fce8f72cfe6383e464abdd8cd86006f281))
+
+## [2.102.0](https://github.com/kiva/ui/compare/v2.101.2...v2.102.0) (2022-04-05)
+
+
+### Features
+
+* deprecate causes homepage experiment and sign up pages ([2fee1ab](https://github.com/kiva/ui/commit/2fee1abd3b33e0e8e8ecea921aa70af347414292))
+
+### [2.101.2](https://github.com/kiva/ui/compare/v2.101.1...v2.101.2) (2022-04-01)
+
+
+### Bug Fixes
+
+* remove duplicate/old data-test attributes ([2061ec5](https://github.com/kiva/ui/commit/2061ec54f2765ab5545b7027e5689feb905c61bb))
+* use data-testid instead of data-test for element attributes ([3a70160](https://github.com/kiva/ui/commit/3a701601eb5b00ecf6ee20f2e9cdc04f4035a282))
+
+### [2.101.1](https://github.com/kiva/ui/compare/v2.101.0...v2.101.1) (2022-04-01)
+
+
+### Bug Fixes
+
+* set node and npm version on ui storybook job ([fab0d76](https://github.com/kiva/ui/commit/fab0d76281b67e1d0182b6f8f99dc1e450e2028b))
+
+## [2.101.0](https://github.com/kiva/ui/compare/v2.100.0...v2.101.0) (2022-04-01)
+
+
+### Features
+
+* remove lunar dependencies and unused graphql mock ([2f28195](https://github.com/kiva/ui/commit/2f28195fd60c30b5eec847298543876067688668))
+* update vue, vue-template-compiler and vue-server-renderer ([aeeff09](https://github.com/kiva/ui/commit/aeeff090c139aba8c79866ef1a39eafa5f32fe29))
+
+
+### Bug Fixes
+
+* update unit test node version ([49ee83f](https://github.com/kiva/ui/commit/49ee83ff0a933de1807f6b66b37f6685836db17e))
+
+## [2.100.0](https://github.com/kiva/ui/compare/v2.99.2...v2.100.0) (2022-03-31)
+
+
+### Features
+
+* automaticallysupportnotice component was created ([48b891e](https://github.com/kiva/ui/commit/48b891ea4de019ff768e7701a021340c79391062))
+* personalized mg page was removed and related components ([b37f1d1](https://github.com/kiva/ui/commit/b37f1d126f9c72859d7c355d8a6063a3a7f4f2df))
+* query and initialization for mg personalization experiment were removed ([2c55c8b](https://github.com/kiva/ui/commit/2c55c8b7775e8436bc08d920bdf6825fd91e974b))
+
+### [2.99.2](https://github.com/kiva/ui/compare/v2.99.1...v2.99.2) (2022-03-30)
+
+
+### Bug Fixes
+
+* ensure custom donation amount is propagated to parent context ([c7ea61f](https://github.com/kiva/ui/commit/c7ea61f17f90e90e873ae26be12adacff5021d33))
+
+### [2.99.1](https://github.com/kiva/ui/compare/v2.99.0...v2.99.1) (2022-03-30)
+
+
+### Bug Fixes
+
+* addtobasket method was updated for a bug adding double donation ([d22b23e](https://github.com/kiva/ui/commit/d22b23ebebf00a7730c79ccb20639a9c9be8163b))
+
 ## [2.99.0](https://github.com/kiva/ui/compare/v2.98.0...v2.99.0) (2022-03-29)